+++ main.go
@@ -16,22 +16,19 @@ package main
 
 import (
 	"context"
-	"database/sql"
 	"flag"
 	"fmt"
-	"google.golang.org/protobuf/encoding/protojson"
+	"io/ioutil"
 	"math/rand"
 	"net/http"
 	"net/url"
 	"os"
 	"os/signal"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"syscall"
 	"time"
-
-	"io/ioutil"
-	"path/filepath"
 
 	"github.com/gofrs/uuid"
 	"github.com/heroiclabs/nakama/v3/ga"
@@ -41,6 +38,7 @@ import (
 	_ "github.com/jackc/pgx/v4/stdlib"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+	"google.golang.org/protobuf/encoding/protojson"
 )
 
 const cookieFilename = ".cookie"
@@ -116,7 +114,7 @@ func main() {
 	}
 	startupLogger.Info("Database connections", zap.Strings("dsns", redactedAddresses))
 
-	db, dbVersion := dbConnect(startupLogger, config)
+	db, dbVersion := server.DbConnect(startupLogger, config)
 	startupLogger.Info("Database information", zap.String("version", dbVersion))
 
 	// Global server context.
@@ -232,55 +230,6 @@ func main() {
 	startupLogger.Info("Shutdown complete")
 
 	os.Exit(0)
-}
-
-func dbConnect(multiLogger *zap.Logger, config server.Config) (*sql.DB, string) {
-	rawURL := config.GetDatabase().Addresses[0]
-	if !(strings.HasPrefix(rawURL, "postgresql://") || strings.HasPrefix(rawURL, "postgres://")) {
-		rawURL = fmt.Sprintf("postgres://%s", rawURL)
-	}
-	parsedURL, err := url.Parse(rawURL)
-	if err != nil {
-		multiLogger.Fatal("Bad database connection URL", zap.Error(err))
-	}
-	query := parsedURL.Query()
-	if len(query.Get("sslmode")) == 0 {
-		query.Set("sslmode", "prefer")
-		parsedURL.RawQuery = query.Encode()
-	}
-
-	if len(parsedURL.User.Username()) < 1 {
-		parsedURL.User = url.User("root")
-	}
-	if len(parsedURL.Path) < 1 {
-		parsedURL.Path = "/nakama"
-	}
-
-	multiLogger.Debug("Complete database connection URL", zap.String("raw_url", parsedURL.String()))
-	db, err := sql.Open("pgx", parsedURL.String())
-	if err != nil {
-		multiLogger.Fatal("Error connecting to database", zap.Error(err))
-	}
-	// Limit the time allowed to ping database and get version to 15 seconds total.
-	ctx, ctxCancelFn := context.WithTimeout(context.Background(), 15*time.Second)
-	defer ctxCancelFn()
-	if err = db.PingContext(ctx); err != nil {
-		if strings.HasSuffix(err.Error(), "does not exist (SQLSTATE 3D000)") {
-			multiLogger.Fatal("Database schema not found, run `nakama migrate up`", zap.Error(err))
-		}
-		multiLogger.Fatal("Error pinging database", zap.Error(err))
-	}
-
-	db.SetConnMaxLifetime(time.Millisecond * time.Duration(config.GetDatabase().ConnMaxLifetimeMs))
-	db.SetMaxOpenConns(config.GetDatabase().MaxOpenConns)
-	db.SetMaxIdleConns(config.GetDatabase().MaxIdleConns)
-
-	var dbVersion string
-	if err = db.QueryRowContext(ctx, "SELECT version()").Scan(&dbVersion); err != nil {
-		multiLogger.Fatal("Error querying database version", zap.Error(err))
-	}
-
-	return db, dbVersion
 }
 
 // Help improve Nakama by sending anonymous usage statistics.