+++ sample_go_module/sample.go
@@ -158,6 +158,14 @@ func (m *Match) MatchTerminate(ctx context.Context, logger runtime.Logger, db *s
 	return state
 }
 
+func (m *Match) MatchSignal(ctx context.Context, logger runtime.Logger, db *sql.DB, nk runtime.NakamaModule, dispatcher runtime.MatchDispatcher, tick int64, state interface{}, data string) (interface{}, string) {
+	if state.(*MatchState).debug {
+		logger.Info("match signal match_id %v tick %v", ctx.Value(runtime.RUNTIME_CTX_MATCH_ID), tick)
+		logger.Info("match signal match_id %v data %v", ctx.Value(runtime.RUNTIME_CTX_MATCH_ID), data)
+	}
+	return state, "signal received: " + data
+}
+
 func eventSessionStart(ctx context.Context, logger runtime.Logger, evt *api.Event) {
 	logger.Info("session start %v %v", ctx, evt)
 }